+++ lib/thawani_payment.dart
@@ -1,4 +1,5 @@
 library thawani_payment;
 
 export 'widgets/pay_button.dart';
-export 'functions.dart';
+export 'pay.dart';
+export 'viewmodel/thawani_customerdelete.dart';