+++ lib/viewmodel/keys_viewmodel.dart
@@ -9,6 +9,7 @@ late String? userSuccessUrl;
 late String? userCancelUrl;
 late String userClintID;
 late String userCustomerID;
+late String userSelectCardLoading;
 bool isTestMode = false;
 late List<Map<String, dynamic>> userProducts;
 late Map<String, dynamic>? userMetadata;